# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Proveedores",
        "description": "Documentación de la API de autenticación y gestión de proveedores.",
        "contact": {
            "name": "Soporte",
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local: SQLite"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Listar productos",
                "operationId": "getProductos",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Producto"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "storeProducto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidInputException"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto con el estado actual del recurso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Obtener un producto específico",
                "operationId": "getProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Producto"
                ],
                "summary": "Actualizar un producto",
                "operationId": "updateProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidInputException"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto con el estado actual del recurso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Producto"
                ],
                "summary": "Eliminar un producto",
                "operationId": "deleteProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "403": {
                        "description": "No se puede eliminar este producto por restricciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteRestrictedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConflictException": {
                "title": "Error de Conflicto (409)",
                "description": "Se lanza cuando hay un conflicto con el estado actual del recurso.",
                "required": [
                    "message",
                    "errorType"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Conflicto con el estado actual del recurso."
                    },
                    "errorType": {
                        "type": "string",
                        "example": "conflict"
                    }
                },
                "type": "object"
            },
            "DeleteRestrictedException": {
                "title": "Error por restricción de eliminación (403)",
                "description": "Se lanza cuando no se puede eliminar un recurso debido a restricciones.",
                "required": [
                    "message",
                    "errorType"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Este recurso no puede eliminarse por restricciones."
                    },
                    "errorType": {
                        "type": "string",
                        "example": "delete_restricted"
                    }
                },
                "type": "object"
            },
            "InvalidInputException": {
                "title": "Entrada inválida (422)",
                "description": "Se lanza cuando la entrada del usuario no cumple con los requisitos.",
                "required": [
                    "message",
                    "errorType"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Entrada inválida."
                    },
                    "errorType": {
                        "type": "string",
                        "example": "invalid_input"
                    }
                },
                "type": "object"
            },
            "ResourceNotFoundException": {
                "title": "Recurso no encontrado (404)",
                "description": "Se lanza cuando el recurso solicitado no existe en el sistema.",
                "required": [
                    "message",
                    "errorType"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado."
                    },
                    "errorType": {
                        "type": "string",
                        "example": "resource_not_found"
                    }
                },
                "type": "object"
            },
...
            "Producto": {
                "required": [
                    "nombre",
                    "proveedor_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Cemento gris 50kg"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Saco de cemento gris para construcción"
                    },
                    "codigo_interno": {
                        "type": "string",
                        "example": "CMG-50"
                    },
                    "precio_unitario": {
                        "type": "number",
                        "format": "float",
                        "example": 180.5
                    },
                    "disponible": {
                        "type": "boolean",
                        "example": true
                    },
                    "proveedor_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "unidad_medida_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "grupo_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}